almost done with calendar

+++ src/app/u/calendar/editCalendarModal.js
@@ -15,7 +15,12 @@ import IconButton from "@mui/material/IconButton";
 import MoreVertRoundedIcon from '@mui/icons-material/MoreVertRounded';
 import DeleteRoundedIcon from '@mui/icons-material/DeleteRounded';
 import CircleRoundedIcon from '@mui/icons-material/CircleRounded';
-
+import List from "@mui/material/List";
+import ListItem from "@mui/material/ListItem";
+import InputAdornment from "@mui/material/InputAdornment";
+import PersonAddRoundedIcon from '@mui/icons-material/PersonAddRounded';
+import AccountCircleRoundedIcon from '@mui/icons-material/AccountCircleRounded';
+import PersonRemoveRoundedIcon from '@mui/icons-material/PersonRemoveRounded';
 
 
 
@@ -26,16 +31,22 @@ import { useRouter } from "next/navigation";
 
 
 
-export default function EditCalendarModal({calendar}) {
+export default function EditCalendarModal({calendarData, teamMembers, user}) {
     const supabase = createClient();
     const router = useRouter();
+    const teamMemberIds = teamMembers?.filter(v=>v.id===user.id).map(v=>v.id) || [];
     const [isPending, startTransition] = useTransition();
     const [open, setOpen] = useState(false);
     const [loading, setLoading] = useState(false);
-    const [nameText, setNameText] = useState(calendar.name);
-    const [colorText, setColorText] = useState(calendar.default_color);
-    const [descriptionText, setDescriptionText] = useState(calendar.description);
+    const [nameText, setNameText] = useState(calendarData.calendars.name);
+    const [colorText, setColorText] = useState(calendarData.calendars.default_color);
+    const [descriptionText, setDescriptionText] = useState(calendarData.calendars.description);
     const fullScreen = useMediaQuery(theme.breakpoints.down('sm'));
+    const [disableType, setDisableType] = useState(false);
+    const [memberText, setMemberText] = useState('');
+    const [errorText, setErrorText] = useState('');
+    const [memberEmails, setMemberEmails] = useState([]);
+    const [memberIds, setMemberIds] = useState([]);
 
     // makes sure that the info is loaded before finishing.
     useEffect(() => {
@@ -44,6 +55,17 @@ export default function EditCalendarModal({calendar}) {
             setLoading(false);
         }
     }, [isPending]);
+
+    // get all member emails and member ids
+    useEffect(() => {
+        async function getProfile() {
+            const {data: profiles, error} = await supabase.rpc('get_calendar_emails', {cid: calendarData.calendars.id})
+            setMemberEmails(profiles.map(p => p.email));
+            setMemberIds(profiles.map(p => p.id))
+        }
+
+        getProfile();
+    }, [calendarData, open]);
 
     // handlers
     async function handleOpen(e) {
@@ -56,8 +78,10 @@ export default function EditCalendarModal({calendar}) {
         e?.stopPropagation();
         // close modal
         setOpen(false);
-        setNameText(calendar.name);
-        setDescriptionText(calendar.description);
+        setNameText(calendarData.calendars.name);
+        setDescriptionText(calendarData.calendars.description);
+        setMemberText('');
+        setErrorText('');
     }
 
     function handleNameText({target}) {
@@ -74,18 +98,87 @@ export default function EditCalendarModal({calendar}) {
         setDescriptionText(target.value);
     }
 
+    function handleMemberText({target}) {
+        // member text
+        setMemberText(target.value);
+        if (errorText.length > 0) setErrorText('');
+    }
+
+    function handleKeyDown(e) {
+        // remove form enter for handling adding members
+        if (e.key === "Enter") {
+            e.preventDefault(); // Prevent form submission
+            handleAddMember();
+        }
+      
+    }
+
+    async function handleAddMember() {
+        // check if member exist then adds it to the list
+        setDisableType(true);
+
+        if (memberEmails.includes(memberText)) {
+            // checks if user is already added
+            setErrorText('User is already added');
+        } else {
+            if (user.email === memberText) {
+                // checks if this is your email
+                setErrorText('This is your email');
+            } else {
+                const { data, error } = await supabase.rpc('email_exists', { checked_email: memberText })
+                if (data[0]) {
+                    // checks if user exists
+                    setMemberEmails(m=>m.concat([memberText]));
+                    setMemberIds(m=>m.concat([data[0].id]));
+                } else {
+                    setErrorText('User does not exist');
+                }
+            }            
+        }
+
+        setDisableType(false);
+        setMemberText('');
+    }
+
+    async function handleRemoveMember({currentTarget}) {
+        // removes a member
+        const value = Number(currentTarget.dataset.value);
+        console.log(calendarData)
+        console.log(memberIds);
+        if (memberEmails[value] === user.email) {
+            // check if you are removing self
+            setErrorText("You can't remove yourself");
+        } else if (calendarData.role !== 'owner') {
+            setErrorText("Only owners can remove members");
+        } else if (teamMemberIds.includes(memberIds[value])) {
+            setErrorText("You can't remove a team member from a team calendar.")
+        } else {
+            // remove
+            setMemberEmails(m => [
+                ...m.slice(0, value),
+                ...m.slice(value+1),
+            ]);
+            setMemberIds(m => [
+                ...m.slice(0, value),
+                ...m.slice(value+1),
+            ]);
+        }
+        
+    }
+
     async function handleSubmit(e) {
         // handle submit
         e.preventDefault();
         setLoading(true);
-        const {data, error} = await supabase
-            .from('calendars')
-            .update({
-                name: nameText,
-                description: descriptionText,
-                default_color: colorText,
-            })
-            .eq('id', calendar.id)
+        const {data, error} = await supabase.rpc('update_calendar', {
+            cid: calendarData.calendars.id,
+            new_name: nameText,
+            new_description: descriptionText,
+            new_default_color: colorText,
+            user_ids: [
+                ...memberIds.map(v => ({user_id: v, role: 'editor'})),
+            ]
+        })
 
         console.log(data);
         console.log(error);
@@ -97,14 +190,21 @@ export default function EditCalendarModal({calendar}) {
 
     async function handleDelete() {
         setLoading(true);
-        const {data, error} = await supabase
-            .from('calendars')
-            .delete()
-            .eq('id', calendar.id);
+        if (calendarData.role === 'owner') {
+            const {data, error} = await supabase
+                .from('calendars')
+                .delete()
+                .eq('id', calendarData.calendars.id);
+            
+            startTransition(() => {
+                router.refresh();
+            })
+        } else {
+            setErrorText('Only owners can delete');
+            setLoading(false);
+        }
 
-        startTransition(() => {
-            router.refresh();
-        })
+        
     }
 
     return (
@@ -121,9 +221,10 @@ export default function EditCalendarModal({calendar}) {
                 onClick={e => e.stopPropagation()}
                 aria-labelledby="alert-dialog-title"
                 aria-describedby="alert-dialog-description"
+                
                 >
                 {/* form */}
-                <form onSubmit={handleSubmit}>
+                <form onSubmit={handleSubmit} sx={{overflow:'hidden'}}>
                     {/* title */}
                     <DialogTitle id="alert-dialog-title" sx={{display:'flex', alignItems: 'center'}}>
                         <Typography variant="inherit">Edit Calendar</Typography>
@@ -173,7 +274,7 @@ export default function EditCalendarModal({calendar}) {
                                     </TextField>
                                 </Box>
                             </Box>
-                            
+                            {/* description */}
                             <TextField 
                                 label='Description'
                                 value={descriptionText}
@@ -186,6 +287,47 @@ export default function EditCalendarModal({calendar}) {
                                     mb:'1rem'
                                 }}
                                 />
+
+                            {/* add users */}
+                            <TextField 
+                                disabled={disableType}
+                                label='Add User'
+                                value={memberText}
+                                onChange={handleMemberText}
+                                onKeyDown={handleKeyDown}
+                                sx={{width:'100%', mb:'1rem'}}
+                                error={errorText}
+                                helperText={errorText}
+                                slotProps={{
+                                    input:{
+                                        endAdornment:(
+                                            <InputAdornment position='end'>
+                                                <IconButton 
+                                                    // disabled={notAdmin} 
+                                                    size='large' edge="end" 
+                                                    onClick={handleAddMember}
+                                                    >
+                                                    <PersonAddRoundedIcon />
+                                                </IconButton>
+                                            </InputAdornment>
+                                        )
+                                    }
+                                }}
+                                />
+                            {/* list of users */}
+                            <Typography variant="h6">Users:</Typography>
+                            <List sx={{maxHeight:'10rem', overflow:'scroll'}}>
+                                {memberEmails.map((m, i) => (
+                                    <ListItem key={i}>
+                                        <AccountCircleRoundedIcon fontSize='large' sx={{mr:'1rem'}} />
+                                        <Typography>{m}</Typography>
+                                        {/* remove user */}
+                                        <IconButton data-value={i} sx={{ml:'auto'}} onClick={handleRemoveMember}>
+                                            <PersonRemoveRoundedIcon fontSize="medium" />
+                                        </IconButton>
+                                    </ListItem>
+                                ))}
+                            </List>
                         </Box>
                     </DialogContent>
                     {/* buttons */}
